requirements.txt
backcall==0.2.0
decorator==4.4.2
Django==2.2
djangorestframework==3.12.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.4
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.3
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
wcwidth==0.2.5
